feat: complete engine hardening and add Go SDK
- engine: segment retention via TTL, WAL rotation on size threshold
- query: SQL-level timeouts and row caps with dedicated per-query runtime
- server: structured API errors with proper HTTP status mapping
- sdk: add Go SDK with slog integration and greplog-go-example
- tests: query API, WAL rotation, retention coverage

+++ crates/engine/src/schema.rs
@@ -3,6 +3,10 @@
 //! Columnar layout mirrors [`crate::record::LogRecord`] one-to-one. Low-cardinality
 //! columns (`level`, `service`) are dictionary-encoded and optional payload columns
 //! (`trace_id`, `raw_body`) are nullable to waste zero space when absent.
+//!
+//! `level` uses an `Int16` key so that even a pathological client cannot exceed
+//! the key space; `MemTable` additionally normalizes every level into a bounded
+//! enumerated set so the dictionary stays tiny in practice.
 
 use std::sync::Arc;
 
@@ -13,7 +17,7 @@ use arrow::datatypes::{DataType, Field, Schema, TimeUnit};
 /// Field order is meaningful:
 /// 1. `timestamp_us` — the partition-pruning axis, microsecond precision.
 /// 2. `trace_id` — nullable correlation id (worker job / HTTP request).
-/// 3. `level` — dictionary-encoded (Int8) severity level.
+/// 3. `level` — dictionary-encoded (Int16) severity level.
 /// 4. `service` — dictionary-encoded (Int16) source application.
 /// 5. `message` — the raw log summary.
 /// 6. `raw_body` — nullable stringified JSON payload.
@@ -24,7 +28,7 @@ pub fn greplog_schema() -> Arc<Schema> {
         Field::new("trace_id", DataType::Utf8, true),
         Field::new(
             "level",
-            DataType::Dictionary(Box::new(DataType::Int8), Box::new(DataType::Utf8)),
+            DataType::Dictionary(Box::new(DataType::Int16), Box::new(DataType::Utf8)),
             false,
         ),
         Field::new(
